# Cleanup of the LFVM hash cache