This PR applies various cleanup steps of the LFVM hash cache implementation. In particular:
- refactoring the hash cache structure to use generic types to eliminate code duplication
- added documentation
- added unit tests for hashing, LRU caching order, and capacity constraints
- added micro-benchmarks to assess impact of code generalization


Benchmarks have been implemented and evaluated to assess the impact of those refactoring steps. The results are as follows:
```
pkg: github.com/Fantom-foundation/Tosca/go/interpreter/lfvm
cpu: Intel(R) Core(TM) i7-6600U CPU @ 2.60GHz
                          │   base.txt   │              mod.txt               │
                          │    sec/op    │   sec/op     vs base               │
HashCache_Hits/size=16-4    517.1n ± 11%   482.8n ± 3%  -6.63% (p=0.001 n=10)
HashCache_Hits/size=32-4    36.73n ±  3%   34.62n ± 1%  -5.76% (p=0.000 n=10)
HashCache_Hits/size=64-4    41.19n ±  2%   39.35n ± 3%  -4.48% (p=0.000 n=10)
HashCache_Hits/size=128-4   527.9n ±  3%   491.1n ± 4%  -6.97% (p=0.000 n=10)
HashCache_Miss/size=16-4    521.5n ±  1%   483.6n ± 1%  -7.27% (p=0.000 n=10)
HashCache_Miss/size=32-4    686.2n ±  2%   644.5n ± 3%  -6.08% (p=0.000 n=10)
HashCache_Miss/size=64-4    740.5n ±  2%   742.3n ± 3%       ~ (p=0.353 n=10)
HashCache_Miss/size=128-4   530.2n ±  1%   497.2n ± 2%  -6.22% (p=0.000 n=10)
geomean                     295.4n         279.4n       -5.42%
```
```
pkg: github.com/Fantom-foundation/Tosca/go/interpreter/lfvm
cpu: Intel(R) Xeon(R) CPU @ 2.80GHz
                           │  base.txt   │               mod.txt               │
                           │   sec/op    │   sec/op     vs base                │
HashCache_Hits/size=16-16    386.2n ± 0%   387.7n ± 2%        ~ (p=0.109 n=10)
HashCache_Hits/size=32-16    25.42n ± 0%   24.22n ± 0%   -4.72% (p=0.000 n=10)
HashCache_Hits/size=64-16    27.74n ± 0%   26.31n ± 6%   -5.16% (p=0.012 n=10)
HashCache_Hits/size=128-16   391.5n ± 0%   395.2n ± 1%   +0.92% (p=0.000 n=10)
HashCache_Miss/size=16-16    386.6n ± 0%   387.4n ± 1%        ~ (p=0.170 n=10)
HashCache_Miss/size=32-16    509.5n ± 1%   523.4n ± 0%   +2.74% (p=0.000 n=10)
HashCache_Miss/size=64-16    549.1n ± 1%   605.6n ± 1%  +10.28% (p=0.000 n=10)
HashCache_Miss/size=128-16   392.7n ± 0%   393.5n ± 0%   +0.22% (p=0.030 n=10)
geomean                      214.8n        215.9n        +0.51%
```