@LuisPH3 2024-09-04
Question, where is the threadsafe requirement coming from? Is this to improve CT or does it play a functional role in the client as well?