@HerbertJordan 2024-09-04
> Question, where is the threadsafe requirement coming from? Is this to improve CT or does it play a functional role in the client as well?

In the client multiple parallel calls to the EVM may be present due to concurrent RPC calls including EVM executions. 

We also have this requirement in CT, Aida EVM validation testing, and Aida EVM + DB validation testing with background queries.